[PH2][Test][Transport] Refactor Chaotic Good Tests
