+++ test/core/transport/chaotic_good/BUILD
@@ -23,10 +23,10 @@ grpc_package(
 )
 
 grpc_cc_library(
-    name = "transport_test",
+    name = "transport_test_helper",
     testonly = 1,
-    srcs = ["transport_test.cc"],
-    hdrs = ["transport_test.h"],
+    srcs = ["transport_test_helper.cc"],
+    hdrs = ["transport_test_helper.h"],
     external_deps = ["gtest"],
     deps = [
         "//:iomgr_timer",
@@ -135,7 +135,7 @@ grpc_cc_test(
     uses_event_engine = False,
     uses_polling = False,
     deps = [
-        "transport_test",
+        "transport_test_helper",
         "//:grpc",
         "//:grpc_public_hdrs",
         "//src/core:arena",
@@ -145,6 +145,7 @@ grpc_cc_test(
         "//src/core:seq",
         "//src/core:slice_buffer",
         "//test/core/transport/util:mock_promise_endpoint",
+        "//test/core/transport/util:transport_test",
     ],
 )
 
@@ -197,7 +198,7 @@ grpc_cc_test(
     uses_event_engine = False,
     uses_polling = False,
     deps = [
-        "transport_test",
+        "transport_test_helper",
         "//:grpc",
         "//:grpc_public_hdrs",
         "//:iomgr_timer",
@@ -212,6 +213,7 @@ grpc_cc_test(
         "//test/core/event_engine/fuzzing_event_engine",
         "//test/core/event_engine/fuzzing_event_engine:fuzzing_event_engine_cc_proto",
         "//test/core/transport/util:mock_promise_endpoint",
+        "//test/core/transport/util:transport_test",
     ],
 )
